HW3/requirements.txt
torch==2.4.0
torchvision==0.19.0
Pillow==10.3.0
tqdm>=4.65
numpy==2.0.1
